refactor: Centralize attribute generator
