+++ tests/bunit.generators.tests/Web.Stub/Components/ButtonComponent.cs
@@ -7,5 +7,6 @@ public class ButtonComponent : ComponentBase
 {
 	[Parameter] public string Text { get; set; }
 	[Parameter] public EventCallback OnClick { get; set; }
+	[CascadingParameter(Name = "Test")] public string Cascading { get; set; }
 	[Required] public string Unused { get; set; }
 }
